Game 2005 A2 v1.2
Third commit, this version has some variables added, they have no functionality as of yet. In addition because you can't see the variables by them self the GUI functionality was added, at least in that you can change the variables. A scale of 1 to 1 is currently the scale due to the smaller numbers of this project

To do
--add simulator functionality